sns_tracer update
