+++ sns_tracer/README.md
@@ -2,3 +2,5 @@
 ```python
 pip install pandas matplotlib youtube-transcript-api transformers kiwipiepy
 ```
+
+`kiwipiepy_model`은 배포 시에 필요할 수 있어, 추가
