+++ sns_tracer/kiwipiepy_model/_version.py
@@ -1,0 +1,1 @@
+__version__ = '0.20.0'
